Declare DisplayInterface and MidiInterface to be used internally
To keep the public API implementation (push.cpp) clean, the midi and display
logic are separated into DisplayInterface and MidiInterface.

The DisplayInterface is responsible for drawing to Push's display, while the
MidiInterface is responsible for sending and receiving midi messages from the
chosen midi port (either the Live port or the User port).