Implement method to draw a frame of 16 bit pixels to Push's display
Frames are passed in as a 2D array of pixels in row-major order.

Public macros are defined so the user can define a statically allocated array
with the same dimensions as the display.

libpush_draw_frame currently blocks until the bulk transfers are complete, but
it shouldn't be too difficult to switch to a non-blocking, multithreaded approach.