Synchronously get reply when sending sysex commands
From a user standpoint it's preferable to have replies from sysex commands
return by the function that sends the command. The issue is that all MIDI
messages are received asynchronously through RtMidi.

To resolve this, add a map from commands to reply queues, i.e. the byte code for
command X maps to the reply queue for command X. When a sysex command is
received, it's arguments are pushed onto the corresponding reply queue in a
thread-safe manner.

The original sysex_call method blocks until the reply queue for its command type
has a reply. Then it gets the reply and returns it.