Add MidiMessageListener template class
In order to avoid duplicating code related to registering and calling callbacks
when midi messages are received, the MidiMessageListener template class is used
with the callback type and the event type as template arguments. A
specialization of the template is instantiated with a pointer to a handler
function that should be used to handle incoming midi messages.

MidiInterface has a specialized listener for each callback type. It sends all
non sysex messages to each of the listeners so that they can handle them appropriately.