Put sysex commands related to leds in LedInterface
Rather than implement the API for all sysex commands in SysexInterface, the define a
separate interface for each type of control. The plan is to provide a single
interface for each type of control (i.e. Pad, Encoder, TouchStrip, RGBButton, WhiteButton) that allows the user to easily access all of
its functionality without worrying about whether it's using sysex messages,
regular midi messages, or usb bulk transfer under the hood. These interfaces can
delegate the details of sending a sysex command and receiving a reply will be delegated to SysexInterface.