Use interface to decouple MidiMessageInterface from handlers
Rather than having MidiInterface contain all of the handlers, the idea is to
instead pass the midi interface to the classes that need to either
send or receive midi messages. These classes can then register message handlers
or send messages.

In this case, SysexInterface needs a reference to MidiInterface in order to send
and receive sysex messages.

The other listeners will soon be removed from MidiInterface and put in their own classes.